ecocharte/requirements.txt
Django==2.2.4
django-bootstrap==0.2.4
django-classy-tags==0.9.0
django-cookie-law==2.0.2
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.2.0
django-fontawesome==1.0
django-haystack==2.8.1
django-leaflet==0.24.0
django-model-utils==3.2.0
django-rest-framework==0.1.0
django-scheduler==0.8.8
django-widget-tweaks==1.4.5
django-summernote==0.8.11.4
requests==2.20.1
icalendar==4.0.3
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
django-simple-captcha==0.5.10
django-widget-tweaks==1.4.3
sqlparse==0.2.4
uni-slugify==0.1.4
Unidecode==1.0.22
urllib3==1.24.2
webencodings==0.5.1
whitenoise==4.1.2
